refactor: move views to mainview
