+++ src/router/index.ts
@@ -1,11 +1,11 @@
 import { createRouter, createWebHashHistory, RouteRecordRaw } from "vue-router";
-import LandingView from '../views/LandingView.vue'
+import MainView from '../views/MainView.vue'
 
 const routes: Array<RouteRecordRaw> = [
   {
     path: '/:currentPage',
-    name: 'LandingView',
-    component: LandingView,
+    name: 'MainView',
+    component: MainView,
   }
 ];
